docker: fix make docker-test and docker-test-html targets
Switching to src/ hierarchy broke the docker-test and docker-test-html
targets, as the old docker setup specifically used the hawkmoth package
from the source, using a mount, and the main point of the src/ hierarchy
was to make this difficult.

Instead, archive the source and expand it in the container, using the
same editable build method. It's not perfect, but it works.