test: add our own src/ directory to the search path
As an artefact of the recent move of the module into the src/ folder,
the unit tests cannot find the local package anylonger. Worse still,
with a sufficiently compatible version of hawkmoth available in the
system path, we may be duped into thinking the tests are working when
they aren't.