parser: workaround cursor.get_tokens() issue with macro expansions
There's a bug in clang that causes cursor.get_tokens() to sometimes fail
when there are macro expansions in the cursor's extent. The result may
be empty or contain bogus tokens. cursor.get_tokens() is unreliable,
unless you can be sure there are no macro expansions.

I haven't been able to come up with a reproducer that would hit this
with our current usage of cursor.get_tokens(). However, it's obvious
when trying to tokenize e.g. 'bool x'. Rather than wait for us hitting
subtle bugs, defensively work around the issue.

The issue seems to be related to something going awry in the cursor's
extent. Simply recreating the extent and getting the tokens from the
translation unit works. The resulting new extent has the same __repr__,
but they differ under the hood. The new one works, the old one doesn't.

This is clang issue [1]. See also [2] and [3].

[1] https://github.com/llvm/llvm-project/issues/68340
[2] https://github.com/llvm/llvm-project/issues/43451
[3] https://stackoverflow.com/questions/16786767/obtain-original-unexpanded-macro-text-using-libclang