extension: add support for source code links
Sphinx has sphinx.ext.viewcode and sphinx.ext.linkcode extensions to
fully incorporate source code into the documentation and to link to
externally hosted source code, respectively. They are not trivial to
integrate with Hawkmoth, however.

Add a dedicated version of linkcode that can be configured with a simple
template URI.