doc: enable source links in examples
This is a bit more convoluted than usual, because we want documentation
built for tagged releases to point at the source code for the release,
and master otherwise. We rely on version having a '.dev0' suffix (or
'.<anything>', really) once we've deviated from the latest release.

The first formatting pass also requires us to wrap {source} and {line}
in extra {}'s.