test: query system include args from clang
Apparently libclang (and consequently Clang Python Bindings) does not
universally get its system include paths right by default, or the same
as clang. The defaults work on Debian, but not on Fedora, for example,
and I've been unable to root cause this.

Anyway, we already have a helper for querying the system include args
from clang, so use it.

Update github workflow while at it, as the examples check now also needs
clang.

Fixes #193.