doc: allow documentation build using hawkmoth from source tree
For the purposes of testing and CI, it's preferrable to use the package
as it is installed instead of directly from source. See commits
9ad923e5cf0f ("build: actually package hawkmoth.ext.*") and f8421b17c821
("move source under src/").

However, sometimes it's convenient to use the extension from the source
tree for the documentation build, instead of forcing the use of a
virtual environment.

Add src/ in Python path when usage of installed package is not
explicitly required and we're not on Read the Docs. Use the
'use-installed-hawkmoth' Sphinx tag for the former, and set it in
SPHINXOPTS by default for 'make html'. The Read the Docs config also
installs the package, so we can use that.

Direct sphinx-build will use the extension from source, as will 'make
SPHINXOPTS="" html'.

Fixes: #203