parser: set clang -x<language> option at the low level
Have a single point of truth for setting the clang -x<language> option,
to unify parsing across the extension, tests, and cli.

Choose the language primarily based on the domain, and secondarily based
on the filename extension to differentiate headers. Using "-header" is
necessary for some cases such as "#pragma once", where plain "-xc" or
"-xc++" would interpret the file as a main file.

This unconditionally bypasses libclang automatic detection. The user and
tests can still override the parser selected language, as the user
provided clang_args are appended.

Fixes: #208