doccursor: move helpers inside the class
Another baby step, which is again mostly cosmetic.

In doing this, we can remove plenty of the attributes we defined just to
get cursor compatibility. We may want to expose some of this again
later, but likely through a better API. For now we simplify the code as
much as possible.

After we should start thinking of how this will play out with docstring
and improving these methods / attributes to provide a nice API for the
new cursor.