doccursor: always use the Clang cursor internally
We want to start improving the API in ways that will no longer mimic the
Clang's cursor one. We need to prevent (e.g.) our cursor `type` property
turns out to mean something different than Clang's own namesake
(spoiler: it will!).

There might be a few cases where the result might be the same, but that
will not be obvious most of the times and it may still pose a
performance issue anyway. As a rule, we stick to using the Clang cursor
everywhere we can.