doccursor: return decl_name as name if name would be None
Clang 15 and earlier have None for the cursor.spelling of typedefs of
anonymous structs/unions/enums. Clang 16 and later return the name of
the typedef. We address this in decl_name to get the Docstring right
across Clang versions, but the name filtering in Docstring._match() uses
name. This results in us being unable to reference typedefs of anonymous
entities with Clang 15 and earlier.

Fix it up by returning decl_name as name from Doccursor if
cursor.spelling is None.

In the long run, we should unify on just name. The decl_name was always
a workaround when we did the fixups in parser.py. With Doccursor, we
should be able to further clean that up, but for the time being, close
the known issue with the minimal fix.

It's probably theoretical, but in the case *both* name and decl_name
were None, we'd end up in infinite recursion if decl_name used self.name
instead of self._cc.spelling. It's a bit of a wart, but should also be
cleaned up later.

Fixes #201