doccursor: cache cursor operations
Some of the methods within the cursors are expensive, so caching them
can have a compounding effect on performance as they are called more
often.

Currently, this patch actually worsens performance a negligible amount
over our unit tests, but in more complicated code bases there's higher
potential for cursors to be revisited often. Even more so when we expose
them to user land over which we have no control naturally.

In order for this to work, we need to make sure that the `__eq__` method
is defined on top of the existing `__hash__`. Together, these guarantee
that we can construct 2 different `DocCursors` from the same arguments
and get to prime the cache from one of them and reuse it in the other
one.

The (maximum) cache size is also made configurable through an
environment variable as the most elegant way of injecting the constant
in the module before it gets evaluated. This may later be exposed to the
user in a configuration variable. 2 functions are also created to dump
the caches' status and wipe them respectively. These too can later be
used to inform the user on how to tune the cache size.

Here we set up caching for all methods that are either callable through
multiple paths or directly exposed through the API, with the exception
of `_specifiers_fixup`. This is to do with the limitation all the method
arguments are hashable, together with the `DocCursor` class itself. As
it turns out, Clang's types are not always hashable, so using a bare
`Type` object as a parameters prevents caching. There are ways around
it, sure, but in this case there's no significant performance impact and
it's not a path that might get exposed to the user as it's only called
through internal methods that can themselves be cached.