build: simplify subdir Makefiles
There are Makefiles in subdirectories to allow running make in the
subdirectories. It's kind of handy, so keep it, but simplify the
Makefiles. Rename 'all' target to 'default' to better describe its
purpose.

Turns out src/hawkmoth/Makefile was not updated after the src/
addition. Fix it while at it.