javadoc: allow space before parameter direction indication
Having whitespace between @param and [direction] fails to match the
param regex:

 @param[in] works
 @param [in] fails

This is detected but not handled properly, leading to a backtrace about
mo.group() being called when mo is None.

Fix the regex and, to an extent, the error handling.

This is just the simplest and quickest fix. There should be better error
handling with proper error messages all around, as well as testing.