extensions: deprecate hawkmoth.ext.transformations and hawkmoth.util.doccompat
The introduction of the cautodoc_transformations config option was a
mistake to begin with. The right way forward is the
hawkmoth-process-docstring event. The hawkmoth.ext.transformations
extension was introduced as a temporary solution to not immediately pull
the plug on cautodoc_transformations.

It's time to deprecate the cautodoc_transformations config, the
hawkmoth.ext.transformations extension, and the hawkmoth.util.doccompat
module.

Everyone should move to either the hawkmoth.ext.javadoc or
hawkmoth.ext.napoleon extensions, or if they're not enough, implement
their own extensions through the hawkmoth-process-docstring event. The
javadoc/doxygen helpers in doccompat are inferior to the
hawkmoth.ext.javadoc extension, and the kernel-doc "support" is two
regexes in a row.

Functionally, the deprecation means 1) not loading
hawkmoth.ext.transformations automatically, i.e. you have to manually
add it to extensions in conf.py, and 2) deprecation messages on loading
the extension and the module.