test: add tests for C++ namespace support
This combines some test input from Stephan Lachnit and myself.

Some of the results are likely not what's desired. Not everything
appears to get the namespace prefix, for example variables.

It's not clear that everything that appears to be missing a namespace
prefix should get a prefix either; for example class/struct/union
members, or enumerators, as they're nested within a parent that does
have the namespace prefix.

More than anything, the test documents the status quo, and provides a
way to make incremental improvements on top.