test: split out _Bool/bool tests to C only
Split out the _Bool/bool tests and run them in C only. They're
problematic in C++ because they depend on stdbool.h, and specifically on
GCC version of it which defines _Bool bool for C++. This is problematic
on Alpine Linux which uses musl, as well as on macOS.

Annoyingly the anonymous struct names also change.