test: combine the asserts for output and errors
With the output and error asserts separate, the test case ends on the
first failing assertion, and we never know what was in errors. But the
clue to the output mismatch might be in the errors. For example, an
include file was not found. By combining the asserts into one, we get
all the differences in the pytest results.