test: define default clang args for testing
For starters, define the language stardard to use. Use c17 and c++17,
fairly arbitrarily. Tests can override this. We can remove the std
option from template.yaml.

This is a bit more convoluted than should be necessary, for a few
reasons:

- Extension tests use the directive options from directive string. For
  that, the defaults should be specified in conf.py, but hawkmoth_clang
  is not domain specific, and can't be used. See [1]. We need to insert
  the defaults in directive options.

- We want to preserve the original directive options, and present those
  in examples documentation. However, in all testing, we need to amend
  the defaults.

Note that the parser selects the -x language option based on domain and
filename suffix.

[1] https://github.com/jnikula/hawkmoth/issues/255