doc: widen alabaster theme page and sidebar
Apparently the defaults for page and sidebar widths are 940 and 220 px,
respectively. Make them slightly wider to allow for more content in the
vertical space, as well as longer entries in the navigation bar without
word wrapping. I'm not entirely happy that they're defined in pixels,
but not going to start fighting the theme.