cfg: add compiler and autoconf options
It's a common problem to want hawkmoth to parse the sources with the
exact headers a given code is meant to be compiled with and doing so
requires some verbosity upfront which can easily be avoided. Provide a
couple of options to do just that.

The defaults are chosen in a technically backwards incompatible way, but
it would take a frankly adversarial case to break it. In fact the new
defaults should yield the default behaviour everyone expected from the
beginning: we all expected libclang's search paths to be an exact match
to clang's, and indeed they are in some distributions.

Finally, the new search paths are added after other user's options since
the headers are searched in a left to right manner, giving the user the
possibility of overriding specific headers despite using this option.

For maximum flexibility, this functionality is gated on the 'autoconf'
settings, which can easily be extended later to support additional
features.