util: make compiler helper language aware
Include paths differ between C and C++ domains (typically C++ has
additional search directories), so we need to take that into account.

Furthermore, it makes sense to specify `-nostdinc` if we are including
the standard include directories explicitly already. Not only we prevent
potential duplication of search paths, we ensure that the resulting
search is strictly confined to the paths the given compiler would search
and not the disjunction of those and libclang's own default search
paths.

Since this is going to be called from within the directives now, it also
makes sense to improve error handling, which we do.