test: fix header search paths in tests
This reverts the effects of commit 1243852, while fixing the
inconsistent handling of C/C++ domains since libclang can't always
figure out the right search paths for its own version of Clang despite
prior claims.

It is still unclear whether the upstream issue is an actual issue or if
this is the intended behaviour
(https://github.com/llvm/llvm-project/issues/18150), though it _is_ an
issue and simply using the fixed helper function and configuration
parameter comes at a suitably low cost to maintenance.

Fixes #262.