ci: examples: install dependencies
The example check script has now an indirect import dependency on
clang's bindings (testenv->hawkmoth.util.compiler) even though it
doesn't actually need them. Fixing it in code implies changing the
module structure, which we care for more than we care for this script
that happens to take advantage of tooling that is otherwise used in
scenarios where the bindings are very much needed (running the tests).