test: optionally select active tests for a test case
Allow each test case to select which of all the supported tests are
active. By default, all of them are, but sometimes it's impossible to
create tests that will pass as they target something specific to those
tests.

The motivation here is the expanding set of configuration variables,
which have not been tested so far, but which only affect extension
tests. Creating a test case for a breaking configuration would then not
work since the parser / cli tests would not see the same breaking
configuration.

Technically, it's unlikely a test case will ever need to be _just_ for
cli or parser, but not extension: anything that breaks those would break
the extension ones. But we do it in a general way anyway for simplicity.