test: allow overriding configuration variables per test
Following the previous commit, we can now specify an 'extension only'
test. This only gets us half the way there though. We now need to be
able to vary the configuration variables.

We could do so by pointing to a whole different conf file, but it's
simpler and neater for now to keep a single standard one and allow
arbitrary overrides instead.