parser: expose the typedef for function pointer typedefs
Sphinx supports two styles of typedef documentation:

.. c:type:: typedef-like declaration
.. c:type:: name

We've opted to use the latter, because, arguably, the point of the
typedef is to hide the underlying type. However, for function pointer
typedefs the information about arguments and return types is
crucial. Use the former style for function pointer typedefs to allow
better documentation of them.

Fixes #274