docker: pass VENV as build arg
Move the venv outside of the workdir (source dir), leaving the workdir
clean of any outside artifacts.

This fixes the reuse compliance check, which normally ignores .venv
based on .gitignore, but that doesn't work when the source dir is not a
git repository, as is the case within the container.