docstring: add DocstringProcessor
Add a base class for doing docstring processing. The class contains a
no-op default process_docstring() method. The goal is to group all
processing together to not have to pass a ton of functions.

Initially, allow passing either process_docstring function or processor
object to get_docstring(), to make migration easier. Once migration is
done, the function will be removed.