trailing-comments: Restructure the comment extraction function
Restructure the _comment_extract function to first determine the token type (i.e.. comment, documentable construct, or something else that either does or doesn't break the comment context), and then extract comments based on the token type that was determined. This just separates the logic for grouping comments from the logic that decides what's an applicable comment.