parser: Check for null cursors from get_tokens
Clang 21's cindex.py should be replacing all returned null cursors with
None. However, due to a bug (as of 21.1.3) it still returns null cursors
from tu.get_tokens.

We need to explicitly check for these because cindex.py 21 also raises
an exception when any property of a null cursor is accessed.

    File "hawkmoth/parser.py", line 169, in _comment_extract
      if token_cursor.kind in [CursorKind.INVALID_FILE,
         ^^^^^^^^^^^^^^^^^
    File "clang/cindex.py", line 1578, in inner
      raise Exception("Tried calling method on a null-cursor.")